Create README.md
This web application was developed as an assignment for the course web design and programing. 
It's an application centered around package delivery. The user is able to register/loing and then he can send his package. He is able to select different delivery companies who are able to deliver the package to the destination address. Anytime during the delivery he is able to track the package location and status. Besides the users there are different functionalities for the moderators as well as for the administrators. 
Application was written in plain php and is using mysql as the storage database. A lot of functionality was developed using ajax and rest services for a smoother user experience